chore: refactor AdyenNativeModules.js
