+++ src/ActionModuleWrapper.tsx
@@ -1,0 +1,22 @@
+import { NativeModule } from 'react-native';
+import { PaymentAction } from './Core/types';
+import { BaseConfiguration } from './Core/configuration';
+import { ActionModule } from './AdyenNativeModules';
+
+export class ActionModuleWrapper implements ActionModule {
+  nativeModule: NativeModule | any;
+  public threeDS2SdkVersion: string;
+
+  constructor(nativeModule: NativeModule | any) {
+    this.nativeModule = nativeModule;
+    this.threeDS2SdkVersion = nativeModule.getConstants().threeDS2SdkVersion;
+  }
+
+  handle(action: PaymentAction, configuration: BaseConfiguration) {
+    return this.nativeModule.handle(action, configuration);
+  }
+
+  hide(success: boolean) {
+    this.nativeModule.hide(success);
+  }
+}
